Move ResponseType impls to package 'responsetypes'
Move ResponseType impls to package 'responsetypes'

[#463]
