+++ core/src/main/java/org/axonframework/queryhandling/responsetypes/ResponseTypes.java
@@ -1,4 +1,6 @@
-package org.axonframework.queryhandling;
+package org.axonframework.queryhandling.responsetypes;
+
+import org.axonframework.queryhandling.ResponseType;
 
 import java.util.List;
